+++ .gitignore
@@ -1,3 +1,3 @@
 /env
 *.DS_Store
-/__pycache__
+*__pycache__
